Clean up standalone_CUPiD page
Since the tutorials are specific to running on Casper, I removed a bunch of
instances of multiple options being given to accomplish the same task (in favor
of our prefered method on the NCAR machines). I also re-organized some of the
tasks and checkpoints.

Also standardized wording of checkpoints and removed trailing whitespace across
the entire diagnostics directory.